"Seperate concerns for photo database"
